Max And Min In Linked List
Given a singly linked list of N elements. The task is to find the maximum and minimum element.

Example 1:
Input:
LinkedList: 1->2->3->4->5
Output: 5 1
Example 2:
Input:
LinkedList: 2->4->6->7->5->1->0
Output: 7 0
Your Task:
Your task is to complete the given functions both of which take head reference as argument and return maximum and minimum element as mentioned. The printing is done by driver code.
Constraints:
1 <= n <= 100
1 <= value <= 103