Search In Linked List
You are given a singly linked list of N elements, and also an element x. You need to find if x is present in the linked list or not.

Example 1:
Input:
LinkedList: 1->2->3->4->5
x = 4
Output: 1
Example 2:
Input:
LinkedList: 2->4->6->7->5->1->0
x = 10
Output: 0
Your Task:
Your task is to complete the given function searchLinkedList() which takes head reference and x as arguments and returns true if x is present else returns false. (1 is printed by the driver code if the returned value is true, otherwise 0)
Constraints:
1 <= n <= 100
1 <= x, value <= 10^3